Added changes for m0betool tracking all btrees

+++ be/tool/main.c
@@ -32,12 +32,29 @@
 
 #include <stdio.h>              /* printf */
 #include <stdlib.h>             /* EXIT_FAILURE */
+#include <string.h>     	/* strcmp */
 
 #include "lib/string.h"         /* m0_streq */
+#include "lib/types.h"          /* PRIu64 */
+#include "lib/misc.h"           /* PRIu64 */
+#include "lib/errno.h"          /* ENOMEM */
 
 #include "be/tool/st.h"         /* m0_betool_st_mkfs */
 #include "be/tool/common.h"     /* m0_betool_m0_init */
 #include "be/ut/helper.h"       /* m0_be_ut_backend */
+#include "be/alloc_internal.h"  /* m0_be_allocator_header */
+#include "stob/ad_private.h"    /* stob_ad_0type_rec */
+#include "cob/cob.h"            /* m0_cob_domain */
+#include "stob/ad.h"            /* m0_stob_ad_domain */
+#include "be/extmap_internal.h" /* m0_be_emap */
+#include "balloc/balloc.h"      /* m0_balloc */
+#include "be/domain.h"	       /*m0_be_segobj_opt_begine,m0_be_segobj_opt_next*/
+#include "lib/uuid.h"		/* m0_node_uuid_string_set */
+
+M0_TL_DESCR_DEFINE(zt, "m0_be_domain::bd_0types", M0_INTERNAL,
+			   struct m0_be_0type, b0_linkage, b0_magic,
+			   M0_BE_0TYPE_MAGIC, M0_BE_0TYPE_MAGIC);
+M0_TL_DEFINE(zt, static, struct m0_be_0type);
 
 
 static const char *betool_help = ""
@@ -46,6 +63,8 @@ static const char *betool_help = ""
 "- 'st mkfs'\n"
 "- 'st run'\n"
 "- 'be_recovery_run'\n"
+"- 'track_btree'\n"
+"- 'print_btree'\n"
 "\n"
 "Use case for 'st mkfs' and 'st run': run 'st mkfs' once to initialise \n"
 "BE data structures, then run 'st run' and kill m0betool process during \n"
@@ -58,6 +77,102 @@ static const char *betool_help = ""
 "'path' parameter is an optional path to BE domain stob domain location.\n"
 "Default BE domain stob domain location is used when this parameter is absent:"
 "\n";
+
+extern void btree_dbg_print(struct m0_be_btree *tree);
+
+
+static void scan_btree(struct m0_be_domain *dom, int print_btree)
+{
+	int                     left;
+	char                    *suffix;
+	struct m0_buf            opt;
+	struct m0_be_seg        *seg;
+	struct m0_be_0type      *objtype;
+	struct m0_cob_domain     *cdom;
+	struct stob_ad_0type_rec *rec ;
+	struct m0_balloc         *m0balloc;
+
+	seg = m0_be_domain_seg0_get(dom);
+	if (seg == NULL)
+	{
+		return;
+	}
+
+	m0_tl_for(zt, &dom->bd_0types, objtype) {
+		for (left = m0_be_segobj_opt_begin(seg, objtype, &opt, &suffix);
+		left > 0 ;
+		left = m0_be_segobj_opt_next(seg, objtype, &opt, &suffix)) {
+
+			M0_LOG(M0_ALWAYS, "object b0_name = '%s' suffix = '%s'"
+						"objtype = %p seg = %p",
+						objtype->b0_name, suffix,
+						objtype, seg );
+
+			if (  strcmp(objtype->b0_name, "M0_BE:COB") == 0)
+			{
+				cdom = *(struct m0_cob_domain**)opt.b_addr;
+				M0_LOG(M0_ALWAYS, "M0_BE:COB "
+						"cd_object_index btree = %p "
+						"cd_namespace btree = %p "
+						"cd_fileattr_basic btree = %p "
+						"cd_fileattr_omg btree = %p "
+						"cd_fileattr_ea btree = %p",
+						&cdom->cd_object_index,
+						&cdom->cd_namespace,
+						&cdom->cd_fileattr_basic,
+						&cdom->cd_fileattr_omg,
+						&cdom->cd_fileattr_ea);
+				if(print_btree)
+				{
+					M0_LOG(M0_ALWAYS, "cd_object_index ");
+					btree_dbg_print(&cdom->cd_object_index);
+					M0_LOG(M0_ALWAYS, "cd_namespace ");
+					btree_dbg_print(&cdom->cd_namespace);
+					M0_LOG(M0_ALWAYS, "cd_fileattr_basic ");
+					btree_dbg_print(&cdom->
+							cd_fileattr_basic);
+
+				}
+			}
+			else if ( strcmp(objtype->b0_name, "M0_BE:AD") == 0)
+			{
+				rec = (struct stob_ad_0type_rec *)opt.b_addr;
+				M0_LOG(M0_ALWAYS, "M0_BE:AD em_mapping = %p",
+				     &rec->sa0_ad_domain->sad_adata.em_mapping);
+				if(print_btree)
+				{
+					M0_LOG(M0_ALWAYS, "em_mapping");
+					btree_dbg_print(&rec->sa0_ad_domain->
+							sad_adata.em_mapping);
+				}
+				m0balloc = container_of(
+					       rec->sa0_ad_domain->sad_ballroom,
+					       struct m0_balloc, cb_ballroom);
+				if (m0balloc != NULL)
+				{
+					M0_LOG(M0_ALWAYS,"M0_BE:AD "
+						"cb_db_group_extents btree= %p "
+						"cb_db_group_desc btree= %p",
+						&m0balloc->cb_db_group_extents,
+						&m0balloc->cb_db_group_desc);
+					if(print_btree)
+					{
+						M0_LOG(M0_ALWAYS, "grp_exts");
+						btree_dbg_print(
+							   &m0balloc->
+							   cb_db_group_extents);
+						M0_LOG(M0_ALWAYS, "grp_dsc");
+						btree_dbg_print(
+							      &m0balloc->
+							      cb_db_group_desc);
+					}
+				}
+			}
+		}
+	} m0_tl_endfor;
+
+}
+
 
 static void be_recovery_run(char *path)
 {
@@ -80,6 +195,35 @@ static void be_recovery_run(char *path)
 	m0_betool_m0_fini();
 }
 
+static void scan(char *path, char *what)
+{
+	struct m0_be_ut_backend        ut_be = {};
+	struct m0_be_domain_cfg        cfg = {};
+	struct m0_be_domain           *dom;
+	char                           location[0x100] = {};
+	int                            rc;
+
+	m0_node_uuid_string_set(NULL);
+	m0_betool_m0_init();
+	m0_be_ut_backend_cfg_default(&cfg);
+	if (path != NULL) {
+		snprintf((char *)&location, ARRAY_SIZE(location),
+			 "linuxstob:./%s", path);
+		cfg.bc_stob_domain_location = (char *)&location;
+		M0_LOG(M0_ALWAYS, "location=%s", cfg.bc_stob_domain_location);
+	}
+	rc = m0_be_ut_backend_init_cfg(&ut_be, &cfg, false);
+	M0_ASSERT_INFO(rc == 0, "rc=%d", rc);
+	dom = &ut_be.but_dom;
+	if (m0_streq(what, "track_btree"))
+		scan_btree(dom, 0);
+	if (m0_streq(what, "print_btree"))
+		scan_btree(dom, 1);
+	m0_be_ut_backend_fini(&ut_be);
+	m0_betool_m0_fini();
+}
+
+
 int main(int argc, char *argv[])
 {
 	struct m0_be_domain_cfg  cfg = {};
@@ -91,6 +235,17 @@ int main(int argc, char *argv[])
 		be_recovery_run(path);
 		return EXIT_SUCCESS;
 	}
+	if (argc > 1 && m0_streq(argv[1], "track_btree")) {
+		path = argc > 2 ? argv[2] : NULL;
+		scan(path, "track_btree");
+		return EXIT_SUCCESS;
+	}
+	if (argc > 1 && m0_streq(argv[1], "print_btree")) {
+		path = argc > 2 ? argv[2] : NULL;
+		scan(path, "print_btree");
+		return EXIT_SUCCESS;
+	}
+
 	if (argc == 3 && m0_streq(argv[1], "st") &&
 	    (m0_streq(argv[2], "mkfs") || m0_streq(argv[2], "run"))) {
 		if (m0_streq(argv[2], "mkfs") == 0)
